+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.8.0 - released 08/03/2019
+ * Add `showDropdownOnFocus` prop that causes the picker to open the dropdown whenever it receives the focus
+ * Fix error where `onBlur` callbacks wouldn't be passed the event
+ * Correct peer dependencies to require React 16.3+ instead of 16.0+ (because we use `React.createRef()`)
+ * Minor improvements to docs
+
 # 1.7.0 - released 07/03/2019
  * Add `autoFocus` prop that causes the picker to be focused when it mounts
  * Fix bug where the picker would not have the focus styling when it had the focus